backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pymongo>=4.6.0
python-dotenv>=1.0.0
requests>=2.28.0
aiohttp>=3.9.0
python-whois>=0.9.0